public student page

+++ S_and_C/src/main/java/com/example/s_and_c/DTO/FormDTO/ReviewDTO.java
@@ -13,6 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 public class ReviewDTO {
     private String studentEmailForCompanyOnly;
-    private int internshipId;
+    private long internshipId;
     private List<String> review;
 }
